Potapov/tg_bot/ver_1.3 rework commands, now user don't need to remember his and his pet IDs, 'cause they made from ChatId. also other minor fixes
